requirements.txt
pyyaml>=6.0
requests>=2.28.1
PyGithub>=1.58.0
checkov>=2.3.301
tfsec>=1.28.1
python-dotenv>=1.0.0
pathlib>=1.0.1
logging>=0.4.9
argparse>=1.4.0
slack-sdk>=3.19.5